Address review comments
- ADR status Proposed -> Accepted; fenced block tagged text
- Scenario "reads as the first attempt" said the opposite of the assertion
  bound to it: the reader deliberately reports no attempt, and the ladder's
  ?? 0 is what makes it the first. Reworded and rebound.
- legacyStagedJobAttempt moved to its own module, so a unit test of a pure
  string parser no longer imports the 2300-line processor, and the eventual
  deletion is an rm rather than a diff against the queue's hottest file.
  Its removal condition is now written down.
- Replaced a tautological assertion (a value <= max that was already 0) with
  the real boundary: maxAttempts reads, maxAttempts + 1 does not.
- toolFrame takes named parameters, per the repo rule for multi-arg functions
- decodeDrop comment named a function that never existed and described the
  mechanism this PR replaced
