+++ langwatch/src/server/app-layer/langy/streaming/__tests__/langyToolCallId.unit.test.ts
@@ -11,7 +11,13 @@ const REAL_ID = "oljdh6z0";
 const SIGNATURE = `Eo4RCosRARFNMg${"AbCd-_09".repeat(360)}`;
 const POLLUTED_ID = `${REAL_ID}_ts_${SIGNATURE}`;
 
-function toolFrame(id: string, phase: "start" | "end" = "start") {
+function toolFrame({
+  id,
+  phase = "start",
+}: {
+  id: string;
+  phase?: "start" | "end";
+}) {
   return { type: "tool", id, name: "bash", phase };
 }
 
@@ -20,14 +26,14 @@ describe("langy tool call id", () => {
     describe("when the frame is parsed at the wire boundary", () => {
       /** @scenario A tool id carrying a thought signature is reduced to the real id */
       it("records the tool call under the id the provider issued", () => {
-        const parsed = langyRelayFrameSchema.parse(toolFrame(POLLUTED_ID));
+        const parsed = langyRelayFrameSchema.parse(toolFrame({ id: POLLUTED_ID }));
 
         expect(parsed).toMatchObject({ type: "tool", id: REAL_ID });
       });
 
       /** @scenario A tool id carrying a thought signature is reduced to the real id */
       it("keeps the signature out of everything it parsed", () => {
-        const parsed = langyRelayFrameSchema.parse(toolFrame(POLLUTED_ID));
+        const parsed = langyRelayFrameSchema.parse(toolFrame({ id: POLLUTED_ID }));
 
         expect(JSON.stringify(parsed)).not.toContain("_ts_");
         expect(JSON.stringify(parsed)).not.toContain(SIGNATURE.slice(0, 32));
@@ -39,11 +45,11 @@ describe("langy tool call id", () => {
     describe("when each frame is parsed", () => {
       /** @scenario A start and an end frame for the same call still pair up */
       it("resolves both to the same tool call id", () => {
-        const start = langyRelayFrameSchema.parse(toolFrame(POLLUTED_ID, "start"));
+        const start = langyRelayFrameSchema.parse(toolFrame({ id: POLLUTED_ID, phase: "start" }));
         // The end frame carries the turn's LATER signature — a different blob
         // for the same call, which is exactly why the raw id could not pair.
         const end = langyRelayFrameSchema.parse(
-          toolFrame(`${REAL_ID}_ts_${"Zz90-_18".repeat(400)}`, "end"),
+          toolFrame({ id: `${REAL_ID}_ts_${"Zz90-_18".repeat(400)}`, phase: "end" }),
         );
 
         expect(start).toMatchObject({ id: REAL_ID });
@@ -57,7 +63,7 @@ describe("langy tool call id", () => {
       /** @scenario An ordinary tool id is left exactly as it is */
       it("records the id unchanged", () => {
         for (const id of ["toolu_01ABCdef", "call_9xYz", "bash-7", REAL_ID]) {
-          expect(langyRelayFrameSchema.parse(toolFrame(id))).toMatchObject({
+          expect(langyRelayFrameSchema.parse(toolFrame({ id }))).toMatchObject({
             id,
           });
         }
@@ -71,7 +77,7 @@ describe("langy tool call id", () => {
       it("records the id unchanged", () => {
         // Too short to be a stapled blob, so it is part of the name.
         expect(
-          langyRelayFrameSchema.parse(toolFrame("run_ts_migrations")),
+          langyRelayFrameSchema.parse(toolFrame({ id: "run_ts_migrations" })),
         ).toMatchObject({ id: "run_ts_migrations" });
       });
 
@@ -79,7 +85,7 @@ describe("langy tool call id", () => {
       it("leaves a long suffix alone when it is not base64url", () => {
         const id = `job_ts_${"a.b.c.d.".repeat(12)}`;
 
-        expect(langyRelayFrameSchema.parse(toolFrame(id))).toMatchObject({ id });
+        expect(langyRelayFrameSchema.parse(toolFrame({ id }))).toMatchObject({ id });
       });
     });
   });
@@ -107,7 +113,7 @@ describe("langy tool call id", () => {
       /** @scenario An absurdly long id is refused as an invalid frame */
       it("rejects the frame rather than storing it", () => {
         const result = langyRelayFrameSchema.safeParse(
-          toolFrame("x".repeat(4000)),
+          toolFrame({ id: "x".repeat(4000) }),
         );
 
         expect(result.success).toBe(false);
@@ -119,7 +125,7 @@ describe("langy tool call id", () => {
     describe("when its start is recorded as a durable milestone", () => {
       /** @scenario A tool call's durable key is built from the normalised id */
       it("builds the event's idempotency key from the normalised id", async () => {
-        const frame = langyRelayFrameSchema.parse(toolFrame(POLLUTED_ID));
+        const frame = langyRelayFrameSchema.parse(toolFrame({ id: POLLUTED_ID }));
         if (frame.type !== "tool") throw new Error("expected a tool frame");
 
         const [event] = await new InitiateToolCallCommand().handle({
