+++ langwatch/src/server/event-sourcing/queues/groupQueue/groupQueue.ts
@@ -57,6 +57,7 @@ import {
   readJobRoutingMeta,
   withJobAttempt,
 } from "./jobEnvelope";
+import { legacyStagedJobAttempt } from "./legacyStagedJobAttempt";
 import {
   gqGroupAttemptReadFailuresTotal,
   gqGroupsBlockedTotal,
@@ -112,41 +113,6 @@ function nonEmptyString(value: unknown): string | undefined {
   return typeof value === "string" && value.length > 0 ? value : undefined;
 }
 
-/**
- * The retry count off a staged job id left over from before ADR-080, when the
- * ladder appended `/r/<n>` to the id and read the count back by counting the
- * segments.
- *
- * READ-ONLY, and only as a last resort. A job part-way up the unreadable-body
- * ladder at deploy time recorded its count nowhere else — that path re-staged
- * the value unmodified and never wrote the group's chain — so without this it
- * would come back with a fresh budget and a fail-safe would reset mid-flight.
- * Nothing writes this shape any more: a legacy id is a value to interpret on
- * the way out, not a format to keep alive.
- */
-export function legacyStagedJobAttempt(stagedJobId: string): number {
-  // Anchored to the TRAILING ladder. The legacy markers were always appended,
-  // so nothing before the tail is one — and an unanchored scan would let a
-  // producer's own event id (which is free-form and forms the base of the
-  // staged id) donate a `/r/<n>` segment and set another job's retry budget.
-  const tail = /((?:\/[rp]\/[^/]+)+)$/.exec(stagedJobId)?.[1] ?? "";
-  let highest = 0;
-  for (const [, digits] of tail.matchAll(/\/r\/(\d+)(?=\/|$)/g)) {
-    const value = Number(digits);
-    // The terminal restage stamped a wall clock under this same marker. Reading
-    // one as an attempt would vault past the budget and discard a job that
-    // still had rungs left, so anything outside the ladder's range is not a
-    // count.
-    if (
-      Number.isInteger(value) &&
-      value > highest &&
-      value <= JOB_RETRY_CONFIG.maxAttempts
-    ) {
-      highest = value;
-    }
-  }
-  return highest;
-}
 
 /**
  * Configuration for the group queue.
